cc26x0: refactor deep sleep & wakeup into power
