+++ chips/cc26x0/src/chip.rs
@@ -1,7 +1,6 @@
 use cortexm3::{self, nvic};
 use cc26xx::gpio;
 use cc26xx::peripheral_interrupts::*;
-use kernel::common::regs::ReadWrite;
 
 const X0_RF_CPE1: u32 = 2;
 const X0_RF_CPE0: u32 = 9;
@@ -12,13 +11,9 @@ use timer;
 use uart;
 use kernel;
 use rtc;
-use prcm;
 use kernel::support;
 use peripherals;
-use aux;
-use aon;
-
-use setup::recharge;
+use power;
 
 #[repr(C)]
 #[derive(Clone, Copy)]
@@ -39,27 +34,10 @@ impl From<u32> for SleepMode {
     }
 }
 
-pub struct SystemControlRegisters {
-    scr: ReadWrite<u32, SystemControl::Register>,
-}
-
-register_bitfields![
-    u32,
-    SystemControl [
-        SLEEP_ON_EXIT   OFFSET(1)   NUMBITS(1) [], // Go to sleep after ISR
-        SLEEP_DEEP      OFFSET(2)   NUMBITS(1) [], // Enable deep sleep
-        SEVONPEND       OFFSET(4)   NUMBITS(1) []  // Wake up on all events (even disabled interrupts)
-    ]
-];
-
 pub struct Cc26x0 {
     mpu: (),
     systick: cortexm3::systick::SysTick,
-    sys_ctrl_regs: *const SystemControlRegisters,
 }
-
-const SYS_CTRL_BASE: u32 = 0xE000ED10;
-const AON_IOC: u32 = 0x4009_4000;
 
 impl Cc26x0 {
     pub unsafe fn new() -> Cc26x0 {
@@ -67,7 +45,6 @@ impl Cc26x0 {
             mpu: (),
             // The systick clocks with 48MHz by default
             systick: cortexm3::systick::SysTick::new_with_calibration(48 * 1000000),
-            sys_ctrl_regs: SYS_CTRL_BASE as *const SystemControlRegisters,
         }
     }
 }
@@ -123,79 +100,26 @@ impl kernel::Chip for Cc26x0 {
     }
 
     fn sleep(&self) {
-        unsafe {
-            let sleep_mode: SleepMode = SleepMode::from(peripherals::M.lowest_sleep_mode());
-            let regs = &*self.sys_ctrl_regs;
+        let sleep_mode: SleepMode = SleepMode::from(unsafe {
+            peripherals::M.lowest_sleep_mode()
+        });
 
-            match sleep_mode {
-                SleepMode::DeepSleep => {
-                    peripherals::M.before_sleep(sleep_mode as u32);
+        match sleep_mode {
+            SleepMode::DeepSleep => unsafe {
+                peripherals::M.before_sleep(sleep_mode as u32);
+                power::prepare_deep_sleep();
+            },
+            _ => ()
+        }
 
-                    // Freeze the ioc by a iolatch
-                    let iolatch: &ReadWrite<u32> = &*((AON_IOC + 0xC) as *const ReadWrite<u32>);
-                    iolatch.set(0x00);
+        unsafe { support::wfi() }
 
-                    // Power down the AUX
-                    aux::AUX_CTL.wakeup_event(aux::WakeupMode::AllowSleep);
-
-                    // Set the ram retention to retain SRAM
-                    aon::AON.mcu_set_ram_retention(true);
-
-                    // Force disable dma & crypto
-                    prcm::force_disable_dma_and_crypto();
-
-                    // Disable all domains except Peripherals & Serial
-                    prcm::Power::disable_domain(prcm::PowerDomain::VIMS);
-                    prcm::Power::disable_domain(prcm::PowerDomain::RFC);
-                    prcm::Power::disable_domain(prcm::PowerDomain::CPU);
-
-                    // Disable JTAG entirely
-                    aon::AON.jtag_set_enabled(false);
-
-                    // Enable power down of the MCU
-                    aon::AON.mcu_power_down();
-
-                    recharge::before_power_down(0);
-
-                    rtc::RTC.sync();
-
-                    // Set the deep sleep bit
-                    regs.scr.modify(SystemControl::SLEEP_DEEP::SET + SystemControl::SEVONPEND::SET);
-                },
-                _ => ()
-            }
-
-            support::wfi();
-
-            match sleep_mode {
-                SleepMode::DeepSleep => {
-                    rtc::RTC.sync();
-
-                    aux::AUX_CTL.wakeup_event(aux::WakeupMode::WakeUp);
-
-                    prcm::release_uldo();
-
-                    prcm::Power::enable_domain(prcm::PowerDomain::VIMS);
-                    prcm::Power::enable_domain(prcm::PowerDomain::CPU);
-
-                    rtc::RTC.sync();
-
-                    let iolatch: &ReadWrite<u32> = &*((AON_IOC + 0xC) as *const ReadWrite<u32>);
-                    iolatch.set(0x01);
-
-                    rtc::RTC.sync();
-
-                    recharge::after_power_down();
-
-                    rtc::RTC.sync();
-
-                    // Clear the deep sleep bit
-                    regs.scr.modify(SystemControl::SLEEP_DEEP::CLEAR);
-
-                    peripherals::M.after_wakeup(sleep_mode as u32);
-                },
-                _ => ()
-            }
+        match sleep_mode {
+            SleepMode::DeepSleep => unsafe {
+                power::prepare_wakeup();
+                peripherals::M.after_wakeup(sleep_mode as u32);
+            },
+            _ => ()
         }
     }
 }
